streamlit_app/requirements.txt
# Streamlit Dashboard Requirements

# Core
streamlit>=1.28.0

# AWS SDK
boto3>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
altair>=5.1.0

# Database
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
hypothesis>=6.82.0
